move pair to hashing
