+++ graphjet-core/src/main/java/com/twitter/graphjet/algorithms/counting/tweet/TweetRecommendationInfo.java
@@ -18,7 +18,7 @@ package com.twitter.graphjet.algorithms.counting.tweet;
 
 import java.util.Map;
 
-import com.twitter.graphjet.algorithms.Pair;
+import com.twitter.graphjet.hashing.Pair;
 import com.twitter.graphjet.algorithms.RecommendationType;
 import com.twitter.graphjet.algorithms.counting.TopSecondDegreeByCountRecommendationInfo;
